+++ info/CHANGELOG.md
@@ -7,6 +7,10 @@
 - Helper function for enumerating all processes and threads.
 - Check for process and thread permissions.
 
+### Changed
+
+- Merge and refactor hive file permission checks (CVE-2021-36934).
+
 ## 2025-01-09
 
 ### Changed